pyproject.toml
[project]
name = "uestc-wifi-autologin"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "requests>=2.33.0",
]
